+++ packages/odata-v2/src/index.ts
@@ -1,3 +1,10 @@
+/* eslint-disable tsdoc/syntax */
+/**
+ * [[include:odata-v2/README.md]]
+ * @packageDocumentation
+ * @module @sap-cloud-sdk/odata-v2
+ */
+
 export {
   filterFunctions,
   replace,